+++ packages/core/src/traced-orchestrator.ts
@@ -7,11 +7,11 @@ import {
   SubscriptionArgs,
 } from 'graphql';
 import { ArbitraryObject, Maybe } from '@envelop/types';
-import { EnvelopOrchestrator } from './orchestrator';
-import { isAsyncIterable } from './utils';
+import { EnvelopOrchestrator } from './orchestrator.js';
+import { isAsyncIterable } from './utils.js';
 
 const getTimestamp =
-  typeof globalThis !== 'undefined' && globalThis?.performance?.now ? () => performance.now() : () => Date.now();
+  typeof globalThis !== 'undefined' && globalThis?.performance?.now ? () => globalThis.performance.now() : () => Date.now();
 
 const measure = () => {
   const start = getTimestamp();