+++ packages/core/test/plugins/use-immediate-introspection.spec.ts
@@ -1,7 +1,7 @@
 import { createTestkit } from '@envelop/testing';
-import { useImmediateIntrospection } from '../../src/plugins/use-immediate-introspection';
-import { useExtendContext } from '../../src/plugins/use-extend-context';
-import { schema } from '../common';
+import { useImmediateIntrospection } from '../../src/plugins/use-immediate-introspection.js';
+import { useExtendContext } from '../../src/plugins/use-extend-context.js';
+import { schema } from '../common.js';
 import { getIntrospectionQuery } from 'graphql';
 
 describe('useImmediateIntrospection', () => {