+++ packages/plugins/fragment-arguments/src/index.ts
@@ -1,8 +1,8 @@
 import type { Plugin } from '@envelop/core';
 import type { ParseOptions } from 'graphql/language/parser';
 import type { Source, DocumentNode } from 'graphql';
-import { FragmentArgumentCompatibleParser } from './extended-parser';
-import { applySelectionSetFragmentArguments } from './utils';
+import { FragmentArgumentCompatibleParser } from './extended-parser.js';
+import { applySelectionSetFragmentArguments } from './utils.js';
 
 export function parseWithFragmentArguments(source: string | Source, options?: ParseOptions): DocumentNode {
   const parser = new FragmentArgumentCompatibleParser(source, options);