+++ packages/plugins/generic-auth/tests/use-generic-auth.spec.ts
@@ -8,7 +8,7 @@ import {
   useGenericAuth,
   ValidateUserFn,
   UnauthenticatedError,
-} from '../src';
+} from '../src/index.js';
 
 type UserType = {
   id: number;