+++ packages/plugins/sentry/src/index.ts
@@ -6,7 +6,7 @@ import {
   OnExecuteDoneHookResultOnNextHook,
 } from '@envelop/core';
 import * as Sentry from '@sentry/node';
-import type { Span } from '@sentry/types';
+import type { Span, TraceparentData } from '@sentry/types';
 import { ExecutionArgs, GraphQLError, Kind, OperationDefinitionNode, print, responsePathAsArray } from 'graphql';
 
 export type SentryPluginOptions = {
@@ -61,6 +61,12 @@ export type SentryPluginOptions = {
    * @default operation's name or "Anonymous Operation" when missing)
    */
   transactionName?: (args: ExecutionArgs) => string;
+  /**
+   * Produces tracing data for Transaction
+   *
+   * @default is empty
+   */
+  traceparentData?: (args: ExecutionArgs) => TraceparentData | undefined;
   /**
    * Produces a "op" (operation) of created Span.
    *
@@ -176,6 +182,7 @@ export const useSentry = (options: SentryPluginOptions = {}): Plugin => {
       const document = print(args.document);
       const opName = args.operationName || rootOperation.name?.value || 'Anonymous Operation';
       const addedTags: Record<string, any> = (options.appendTags && options.appendTags(args)) || {};
+      const traceparentData = (options.traceparentData && options.traceparentData(args)) || {};
 
       const transactionName = options.transactionName ? options.transactionName(args) : opName;
       const op = options.operationName ? options.operationName(args) : 'execute';
@@ -192,7 +199,16 @@ export const useSentry = (options: SentryPluginOptions = {}): Plugin => {
           name: transactionName,
           op,
           tags,
+          ...(traceparentData || {}),
         });
+
+        if (!rootSpan) {
+          const error = [
+            `Could not create the root Sentry transaction for the GraphQL operation "${transactionName}".`,
+            `It's very likely that this is because you have not included the Sentry tracing SDK in your app's runtime before handling the request.`,
+          ];
+          throw new Error(error.join('\n'));
+        }
       } else {
         const scope = Sentry.getCurrentHub().getScope();
         const parentSpan = scope?.getSpan();