+++ packages/types/src/hooks.ts
@@ -10,16 +10,10 @@ import type {
   SubscriptionArgs,
   ValidationRule,
 } from 'graphql';
-import { Maybe, PromiseOrValue, AsyncIterableIteratorOrValue } from './utils';
-import { DefaultContext } from './context-types';
-import {
-  ExecuteFunction,
-  ParseFunction,
-  ValidateFunction,
-  ValidateFunctionParameter,
-  SubscribeFunction,
-} from './graphql';
-import { Plugin } from './plugin';
+import { Maybe, PromiseOrValue, AsyncIterableIteratorOrValue } from './utils.js';
+import { DefaultContext } from './context-types.js';
+import { ExecuteFunction, ParseFunction, ValidateFunction, ValidateFunctionParameter, SubscribeFunction } from './graphql.js';
+import { Plugin } from './plugin.js';
 
 export type DefaultArgs = Record<string, unknown>;