+++ packages/plugins/newrelic/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @envelop/newrelic
 
+## 4.1.0
+
+### Minor Changes
+
+- 8bb2738: Support TypeScript module resolution.
+- Updated dependencies [8bb2738]
+  - @envelop/core@2.4.0
+
 ## 4.0.2
 
 ### Patch Changes