+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @envelop/testing
 
+## 4.4.0
+
+### Minor Changes
+
+- 8bb2738: Support TypeScript module resolution.
+- Updated dependencies [8bb2738]
+  - @envelop/core@2.4.0
+  - @envelop/types@2.3.0
+
 ## 4.3.3
 
 ### Patch Changes